fix(mobile): address CRAP gate review
- fix widget in-flight cleanup

- share CRAP scope config

- split auth sign-in helpers
