+++ apps/mobile/crap-scope.mjs
@@ -1,0 +1,32 @@
+export const CRAP_SRC = [
+  "features/account-suggestions",
+  "features/activity",
+  "features/advisor",
+  "features/analytics",
+  "features/auth",
+  "features/backup",
+  "features/budget",
+  "features/calendar",
+  "features/capture-evidence",
+  "features/capture-interpreter",
+  "features/capture-sources",
+  "features/categories",
+  "features/dashboard",
+  "features/financial-accounts",
+  "features/goals",
+  "features/onboarding",
+  "features/qa",
+  "features/review-queues",
+  "features/search",
+  "features/settings",
+  "features/transactions",
+  "features/transfers",
+  "shared/constants",
+  "shared/hooks",
+  "shared/i18n",
+  "shared/lib",
+  "shared/mutations",
+  "shared/types",
+];
+
+export const CRAP_EXCLUDE = ["**/components/**", "**/ui/**", "**/*.tsx"];
