+++ apps/mobile/crap4ts.config.ts
@@ -1,38 +1,10 @@
 import { defineConfig } from "crap4ts";
+import { CRAP_EXCLUDE, CRAP_SRC } from "./crap-scope.mjs";
 
 export default defineConfig({
   threshold: 8,
   coverageMetric: "line",
   top: 30,
-  src: [
-    "features/account-suggestions",
-    "features/activity",
-    "features/advisor",
-    "features/analytics",
-    "features/auth",
-    "features/backup",
-    "features/budget",
-    "features/calendar",
-    "features/capture-evidence",
-    "features/capture-interpreter",
-    "features/capture-sources",
-    "features/categories",
-    "features/dashboard",
-    "features/financial-accounts",
-    "features/goals",
-    "features/onboarding",
-    "features/qa",
-    "features/review-queues",
-    "features/search",
-    "features/settings",
-    "features/transactions",
-    "features/transfers",
-    "shared/constants",
-    "shared/hooks",
-    "shared/i18n",
-    "shared/lib",
-    "shared/mutations",
-    "shared/types",
-  ],
-  exclude: ["**/components/**", "**/ui/**", "**/*.tsx"],
+  src: CRAP_SRC,
+  exclude: CRAP_EXCLUDE,
 });
