+++ apps/mobile/vitest.crap.config.ts
@@ -1,4 +1,5 @@
 import baseConfig from "./vitest.config";
+import { CRAP_EXCLUDE, CRAP_SRC } from "./crap-scope.mjs";
 
 export default {
   ...baseConfig,
@@ -6,42 +7,8 @@ export default {
     ...baseConfig.test,
     coverage: {
       ...baseConfig.test?.coverage,
-      include: [
-        "features/account-suggestions/**/*.ts",
-        "features/activity/**/*.ts",
-        "features/advisor/**/*.ts",
-        "features/analytics/**/*.ts",
-        "features/auth/**/*.ts",
-        "features/backup/**/*.ts",
-        "features/budget/**/*.ts",
-        "features/calendar/**/*.ts",
-        "features/capture-evidence/**/*.ts",
-        "features/capture-interpreter/**/*.ts",
-        "features/capture-sources/**/*.ts",
-        "features/categories/**/*.ts",
-        "features/dashboard/**/*.ts",
-        "features/financial-accounts/**/*.ts",
-        "features/goals/**/*.ts",
-        "features/onboarding/**/*.ts",
-        "features/qa/**/*.ts",
-        "features/review-queues/**/*.ts",
-        "features/search/**/*.ts",
-        "features/settings/**/*.ts",
-        "features/transactions/**/*.ts",
-        "features/transfers/**/*.ts",
-        "shared/constants/**/*.ts",
-        "shared/hooks/**/*.ts",
-        "shared/i18n/**/*.ts",
-        "shared/lib/**/*.ts",
-        "shared/mutations/**/*.ts",
-        "shared/types/**/*.ts",
-      ],
-      exclude: [
-        ...(baseConfig.test?.coverage?.exclude ?? []),
-        "**/components/**",
-        "**/ui/**",
-        "**/*.tsx",
-      ],
+      include: CRAP_SRC.map((sourceDir) => `${sourceDir}/**/*.ts`),
+      exclude: [...(baseConfig.test?.coverage?.exclude ?? []), ...CRAP_EXCLUDE],
       thresholds: undefined,
     },
   },
